# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_api.CreateLinkRequest:
    properties:
      description:
        type: string
      expires_at:
        type: string
      slug:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      url:
        type: string
      visibility:
        type: string
    type: object
...
  internal_api.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
    type: object
  internal_api.HealthResponse:
    properties:
      last_checked_at:
        type: string
      last_status:
        type: integer
      status:
        enum:
        - unchecked
        - ok
        - broken
        - skipped
        type: string
    type: object
  internal_api.LinkListResponse:
    properties:
      links:
        items:
          $ref: '#/definitions/internal_api.LinkResponse'
        type: array
      next_cursor:
        type: string
    type: object
  internal_api.LinkResponse:
    properties:
      archived_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      expires_at:
        type: string
      health:
        $ref: '#/definitions/internal_api.HealthResponse'
      id:
        type: string
      lifecycle_state:
        enum:
        - active
        - expired
        - archived
        type: string
      owners:
        items:
          $ref: '#/definitions/internal_api.OwnerResponse'
        type: array
      slug:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      url:
        type: string
      visibility:
        type: string
    type: object
...
  internal_api.OwnerResponse:
    properties:
      email:
        type: string
      id:
        type: string
      is_primary:
        type: boolean
    type: object
...
  internal_api.TagListResponse:
    properties:
      next_cursor:
        type: string
      tags:
        items:
          $ref: '#/definitions/internal_api.TagResponse'
        type: array
    type: object
  internal_api.TagResponse:
    properties:
      link_count:
        type: integer
      name:
        type: string
      slug:
        type: string
    type: object
...
  internal_api.UpdateRoleRequest:
    properties:
      role:
        type: string
    type: object
  internal_api.UserListResponse:
    properties:
      next_cursor:
        type: string
      users:
        items:
          $ref: '#/definitions/internal_api.UserResponse'
        type: array
    type: object
  internal_api.UserResponse:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      role:
        type: string
    type: object
info:
  contact: {}
  description: Self-hosted go-links service. Authenticate with a Personal Access Token.
  title: joe-links API
  version: "1.0"
paths:
  /admin/links:
    get:
      consumes:
      - application/json
      description: Returns all links system-wide with owners and tags. Requires admin
        role.
      parameters:
      - description: Max items to return (default 50, max 200)
        in: query
        name: limit
        type: integer
      - description: Opaque pagination cursor from a prior next_cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.LinkListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - BearerToken: []
      summary: List all links (admin)
      tags:
      - Admin
  /admin/users:
    get:
      consumes:
      - application/json
      description: Returns all users in the system. Requires admin role.
      parameters:
      - description: Max items to return (default 50, max 200)
        in: query
        name: limit
        type: integer
      - description: Opaque pagination cursor from a prior next_cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.UserListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - BearerToken: []
      summary: List all users (admin)
      tags:
      - Admin
  /admin/users/{id}/role:
    put:
      consumes:
      - application/json
      description: 'Changes a user''s role. Valid values: "user", "admin". Requires
        admin role.'
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: New role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - BearerToken: []
      summary: Update user role (admin)
      tags:
      - Admin
...
  /links:
    get:
      consumes:
      - application/json
      description: Returns links owned by the caller. Admins see all links.
      parameters:
      - description: Max items to return (default 50, max 200)
        in: query
        name: limit
        type: integer
      - description: Opaque pagination cursor from a prior next_cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.LinkListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - BearerToken: []
      summary: List links
      tags:
      - Links
    post:
      consumes:
      - application/json
      description: Creates a new short link. The caller becomes the primary owner.
      parameters:
      - description: Link to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.CreateLinkRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api.LinkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - BearerToken: []
      summary: Create a link
      tags:
      - Links
...
  /tags:
    get:
      consumes:
      - application/json
      description: Returns tags that have at least one associated link visible to
        the caller (public, owned, co-owned, or shared). Link counts include only
        visible links. Admins see all tags with full counts.
      parameters:
      - description: Max items to return (default 50, max 200)
        in: query
        name: limit
        type: integer
      - description: Opaque pagination cursor from a prior next_cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.TagListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - BearerToken: []
      summary: List tags
      tags:
      - Tags
  /tags/{slug}/links:
    get:
      consumes:
      - application/json
      description: Returns links with the given tag. Admins see all; non-admins see
        only links visible to them (public, owned, co-owned, or shared). Returns 404
        when the tag does not exist — or, for non-admins, when it has no visible links.
      parameters:
      - description: Tag slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.LinkListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - BearerToken: []
      summary: List links by tag
      tags:
      - Tags
...
securityDefinitions:
  BearerToken:
    description: 'Type "Bearer" followed by a space and your API token. Example: "Bearer
      jl_xxx"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"